Server running
app.js is communicating with server.js, and express and nodemon are installed. At this point I have 2 test routes. One that displays "Hello from express" in the browser, and another that log's "First middleware" in the console. You can view this by initializing the app with nodemon and going to localhost:3000 in the browser.