mongo db hooked up to the back end
As of right now the database has a list of "posts" in it (I'll add users in as well when we get everything hooked to the front end. The database succesfully connects to port 3000 via express (went back and used express generator), and if you travel to /api/posts you can see the list of posts from the database. I also added the body-parser and cors to the app.js.